Added Detail Page and Data Indexs
Added a page to display the data from each specrun, also added four fields, for each datapoint in the tuple, a field in the database tells us what it represents. Not sure if thats the best way to define what each thing means in the tuple but for now we have the information and can use it when needed.